local_side/requirements.txt
opencv-python 
numpy
farm-ng-amiga
pyserial
websockets
pynmea2
pygame